double -> int
Former-commit-id: 0d844b82e4a1f8f5d3ef8101a338de742509a2d3

+++ ide/src/org/alice/ide/croquet/models/project/TreeNodesAndManagers/SearchObject.java
@@ -163,7 +163,7 @@ public class SearchObject<T extends AbstractDeclaration> {
 		return root;
 	}
 
-	public double getReferenceCount() {
+	public int getReferenceCount() {
 		return references.size();
 	}
 }
